# docs/openapi/flagman-delivery.openapi.yaml
openapi: 3.0.3
info:
  title: Flagman Delivery API
  version: 1.0.0
  description: |
    OpenAPI specification generated from the FastAPI backend in `hac.new` and normalized for iOS client generation.
    The backend exposes JWT bearer authentication via `/api/auth/login` and `/api/auth/refresh`.
    Password reset and email verification flows are also described so the spec can be used for iOS client generation.
servers:
  - url: http://localhost:8000
    description: Local development
...
security:
  - bearerAuth: []
paths:
  /api/auth/register:
    post:
      tags: [auth]
      operationId: registerUser
      security: []
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
            example:
              email: customer@example.com
              password: password123
              full_name: Ivan Petrov
              role: customer
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRegisterResponse'
              example:
                id: 101
                email: customer@example.com
                full_name: Ivan Petrov
                role: customer
                message: Регистрация успешна
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/auth/login:
    post:
      tags: [auth]
      operationId: loginUser
      security: []
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
            example:
              username: customer@example.com
              password: password123
      responses:
        '200':
          description: JWT access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
              example:
                access_token: eyJhbGciOi...
                refresh_token: eyJhbGciOi...refresh
                token_type: bearer
                role: customer
                expires_in: 1800
                refresh_expires_in: 604800
                is_verified: false
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/auth/refresh:
    post:
      tags: [auth]
      operationId: refreshAuthToken
      security: []
      summary: Refresh access token pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
            example:
              refresh_token: opaque-or-jwt-refresh-token
      responses:
        '200':
          description: Refreshed token pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/auth/logout:
    post:
      tags: [auth]
      operationId: logoutUser
      summary: Logout current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
            example:
              refresh_token: opaque-or-jwt-refresh-token
      responses:
        '200':
          description: Logout accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionMessageResponse'
              example:
                message: Session closed
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/me:
    get:
      tags: [auth]
      operationId: getCurrentUser
      summary: Get current authenticated user with role profile
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMeResponse'
              example:
                id: 101
                email: customer@example.com
                full_name: Ivan Petrov
                role: customer
                is_verified: true
                profile:
                  id: 15
                  user_id: 101
                  phone: "+79998887766"
                  default_address: "Москва, Арбат 10"
                  default_coordinates:
                    lat: 55.7525
                    lon: 37.5929
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/auth/forgot-password:
    post:
      tags: [auth]
      operationId: requestPasswordReset
      security: []
      summary: Request password reset link or token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
            example:
              email: customer@example.com
      responses:
        '200':
          description: Reset flow started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionMessageResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/auth/reset-password:
    post:
      tags: [auth]
      operationId: confirmPasswordReset
      security: []
      summary: Confirm password reset with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordConfirmRequest'
            example:
              token: reset-token
              new_password: new-password-123
      responses:
        '200':
          description: Password updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionMessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/auth/verify-email/request:
    post:
      tags: [auth]
      operationId: requestEmailVerification
      summary: Send or regenerate email verification token
      responses:
        '200':
          description: Verification flow started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionMessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/verify-email/confirm:
    post:
      tags: [auth]
      operationId: confirmEmailVerification
      security: []
      summary: Confirm email with verification token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerificationConfirmRequest'
            example:
              token: verify-token
      responses:
        '200':
          description: Email verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionMessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: Адрес доставки вне зоны доставки магазина
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: Недействительный или просроченный токен
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: Заказ не найден
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: Заказ недоступен или уже назначен
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
    ValidationIssue:
      type: object
      required: [loc, msg, type]
      properties:
        loc:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
        msg:
          type: string
        type:
          type: string
    ValidationErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
    UserRole:
      type: string
      enum: [customer, courier, business, admin]
...
    UserRegister:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
        full_name:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/UserRole'
    UserLogin:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    LogoutRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    ActionMessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordConfirmRequest:
      type: object
      required: [token, new_password]
      properties:
        token:
          type: string
        new_password:
          type: string
          minLength: 8
          maxLength: 128
    EmailVerificationConfirmRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
    TokenResponse:
      type: object
      required: [access_token, refresh_token, token_type, role, expires_in, refresh_expires_in, is_verified]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        expires_in:
          type: integer
        refresh_expires_in:
          type: integer
        is_verified:
          type: boolean
    UserRegisterResponse:
      type: object
      required: [id, email, role, message]
      properties:
        id:
          type: integer
        email:
          type: string
        full_name:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/UserRole'
        message:
          type: string
...
    UserMeResponse:
      type: object
      required: [id, email, role, is_verified]
      properties:
        id:
          type: integer
        email:
          type: string
        full_name:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/UserRole'
        is_verified:
          type: boolean
        profile:
          type: object
          nullable: true
          description: Generator-friendly superset for business/courier/customer profile payload.
...